+++ backend/src/main/java/cz/vse/swoe/ontodeside/patomat2/model/ResultBinding.java
@@ -1,4 +1,8 @@
 package cz.vse.swoe.ontodeside.patomat2.model;
 
-public record ResultBinding(String name, String value, String datatype) {
+public record ResultBinding(String name, String value, String datatype, boolean basedOnBlankNode) {
+
+    public ResultBinding(String name, String value, String datatype) {
+        this(name, value, datatype, false);
+    }
 }