+++ backend/src/main/java/cz/vse/swoe/ontodeside/patomat2/service/TransformationService.java
@@ -76,7 +76,7 @@ public class TransformationService {
     }
 
     private void applyTransformation(PatternInstance instance, boolean applyDeletes) {
-        if (applyDeletes) {
+        if (applyDeletes && instance.hasDelete()) {
             ontologyHolder.applyTransformationQuery(instance.sparqlDelete());
         }
         ontologyHolder.applyTransformationQuery(instance.sparqlInsert());