+++ backend/src/test/java/cz/vse/swoe/ontodeside/patomat2/model/PatternTest.java
@@ -9,6 +9,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
 @ExtendWith(MockitoExtension.class)
 class PatternTest {
@@ -108,6 +109,23 @@ class PatternTest {
                 "}", sut.createTargetDeleteSparql(new PatternMatch(sut, List.of(
                 new ResultBinding("p", p, Constants.RDFS_RESOURCE),
                 new ResultBinding("A", a, Constants.RDFS_RESOURCE),
+                new ResultBinding("B", b, Constants.RDFS_RESOURCE),
+                new ResultBinding("C", c, Constants.RDFS_RESOURCE)
+        ))));
+    }
+
+    @Test
+    void createTargetDeleteSparqlReturnsNullWhenPatternMatchContainsBlankNodeBasedBinding() {
+        final Pattern sut = new Pattern("filename.json", "name", List.of("?p rdfs:domain ?A",
+                "?p rdfs:range ?B",
+                "?C rdfs:subClassOf ?B"), List.of(), List.of(), List.of());
+        final String p = Generator.generateUri().toString();
+        final String a = Generator.generateUri().toString();
+        final String b = Generator.generateUri().toString();
+        final String c = Generator.generateUri().toString();
+        assertNull(sut.createTargetDeleteSparql(new PatternMatch(sut, List.of(
+                new ResultBinding("p", p, Constants.RDFS_RESOURCE),
+                new ResultBinding("A", a, Constants.RDFS_RESOURCE, true),
                 new ResultBinding("B", b, Constants.RDFS_RESOURCE),
                 new ResultBinding("C", c, Constants.RDFS_RESOURCE)
         ))));