+++ backend/src/test/java/cz/vse/swoe/ontodeside/patomat2/service/rdf4j/Rdf4jOntologyHolderTest.java
@@ -1,5 +1,7 @@
 package cz.vse.swoe.ontodeside.patomat2.service.rdf4j;
 
+import cz.vse.swoe.ontodeside.patomat2.model.Pattern;
+import cz.vse.swoe.ontodeside.patomat2.model.PatternMatch;
 import org.eclipse.rdf4j.model.Model;
 import org.eclipse.rdf4j.model.util.Models;
 import org.eclipse.rdf4j.model.vocabulary.SKOS;
@@ -11,9 +13,11 @@ import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
+import java.util.List;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -63,5 +67,28 @@ class Rdf4jOntologyHolderTest {
         final Model expectedModel = Rio.parse(new FileInputStream(file), RDFFormat.NTRIPLES);
         final Model actualModel = Rio.parse(new ByteArrayInputStream(result.toByteArray()), RDFFormat.NTRIPLES);
         assertTrue(Models.isomorphic(expectedModel, actualModel));
+    }
+
+    @Test
+    void findMatchesReplacesBlankNodeRepresentingUnionWithUnionConstituents() {
+        final File file = new File(getClass().getClassLoader().getResource("cmt.owl").getFile());
+        final Rdf4jOntologyHolder sut = new Rdf4jOntologyHolder();
+        sut.loadOntology(file);
+        final Pattern p = new Pattern("pattern.json", "Pattern", List.of("?p rdfs:domain ?A",
+                "?p rdfs:range ?B",
+                "?C rdfs:subClassOf ?B"), List.of(), List.of(), List.of());
+
+        final List<PatternMatch> result = sut.findMatches(p);
+        final List<PatternMatch> blankNodeOnes = result.stream().filter(PatternMatch::isBasedOnBlankNode).toList();
+        assertFalse(blankNodeOnes.isEmpty());
+        assertTrue(blankNodeOnes.stream()
+                                .anyMatch(pm -> pm.getBinding("A").isPresent() && pm.getBinding("A").get().value()
+                                                                                    .equals("http://cmt/#Reviewer")));
+        assertTrue(blankNodeOnes.stream()
+                                .anyMatch(pm -> pm.getBinding("A").isPresent() && pm.getBinding("A").get().value()
+                                                                                    .equals("http://cmt/#Chairman")));
+        assertTrue(blankNodeOnes.stream()
+                                .anyMatch(pm -> pm.getBinding("A").isPresent() && pm.getBinding("A").get().value()
+                                                                                    .equals("http://cmt/#Author")));
     }
 }